# docs/swagger.yaml
definitions:
...
  models.BNPL:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      customer_id:
        type: string
      id:
        type: string
      paid_amount:
        type: integer
      products:
        additionalProperties:
          $ref: "#/definitions/models.SalesSessionItem"
        description: products in the BNPL
        type: object
      status:
        allOf:
          - $ref: "#/definitions/models.BNPLStatus"
        description: active, completed, cancelled
      total_amount:
        type: integer
      transactions:
        description: id of transactions
        items:
          type: string
        type: array
      updated_at:
        type: string
    type: object
  models.BNPLStatus:
    enum:
      - active
      - completed
      - cancelled
    type: string
    x-enum-varnames:
      - BNPLStatusActive
      - BNPLStatusCompleted
      - BNPLStatusCancelled
...
  models.Error:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.NewBNPLInput:
    properties:
      branch_id:
        type: string
      calculate_total_amount:
        type: boolean
      customer_id:
        type: string
      products:
        additionalProperties:
          $ref: "#/definitions/models.SalesSessionItem"
        type: object
      total_amount:
        type: integer
    type: object
...
  models.Output:
    properties:
      data: {}
      error:
        items:
          $ref: "#/definitions/models.Error"
        type: array
    type: object
...
  models.SalesSessionItem:
    properties:
      price:
        type: integer
      quantity:
        type: integer
    type: object
...
info:
  contact:
    email: hamidovaslon13@gmail.com
    name: API Support
    url: https://github.com/aslon1213/g4h_pos_erp
  description: This is a ERP/POS API for Magazin.
  title: Magazin ERP/POS API
  version: "1.0"
paths:
...
  /api/bnpl:
    post:
      consumes:
        - application/json
      description: Create a new Buy Now Pay Later transaction
      parameters:
        - description: BNPL input
          in: body
          name: input
          required: true
          schema:
            $ref: "#/definitions/models.NewBNPLInput"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/models.Error"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Create new BNPL
      tags:
        - BNPL
  /api/bnpl/{id}:
    delete:
      consumes:
        - application/json
      description: Delete an existing BNPL transaction
      parameters:
        - description: BNPL ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Delete BNPL
      tags:
        - BNPL
    get:
      consumes:
        - application/json
      description: Get details of a specific BNPL transaction
      parameters:
        - description: BNPL ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Get BNPL details
      tags:
        - BNPL
  /api/bnpl/{id}/credit:
    post:
      consumes:
        - application/json
      description: Add a credit payment to an existing BNPL
      parameters:
        - description: BNPL ID
          in: path
          name: id
          required: true
          type: string
        - description: Payment amount
          in: query
          name: amount
          required: true
          type: integer
        - default: cash
          description: Payment method
          in: query
          name: payment_method
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.BNPL"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Credit BNPL payment
      tags:
        - BNPL
  /api/branches/{branch_id}/bnpls:
    get:
      consumes:
        - application/json
      description: Get all BNPL transactions for a specific branch
      parameters:
        - description: Branch ID
          in: path
          name: branch_id
          required: true
          type: string
        - description: Customer name
          in: query
          name: customer_name
          type: string
        - description: Customer phone
          in: query
          name: customer_phone
          type: string
        - description: Customer address
          in: query
          name: customer_address
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Get BNPLs of branch
      tags:
        - BNPL
...
  /api/customers/{customer_id}/bnpls:
    get:
      consumes:
        - application/json
      description: Get all BNPL transactions for a specific customer
      parameters:
        - description: Customer ID
          in: path
          name: customer_id
          required: true
          type: string
        - description: Branch ID
          in: query
          name: branch_id
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Output"
        "500":
          description: Internal Server Error
          schema:
            $ref: "#/definitions/models.Error"
      security:
        - BearerAuth: []
      summary: Get customer BNPLs
      tags:
        - BNPL
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"